add header for simple world
