+++ src/inc/garnet/GNrdg2.h
@@ -4,7 +4,7 @@
 ///
 namespace GN::rdg2 {};
 
-#include "GNgpu2.h"
+#include "GNengine2.h"
 
 // rdg2 subheaders below must only be included through this file (__GN_INSIDE_RDG2_H__ is checked there).
 #define __GN_INSIDE_RDG2_H__ 1
