+++ src/inc/garnet/e2/thing.h
@@ -28,33 +28,6 @@ using ArrayView = SafeArrayAccessor<T>;
 /// Unit of time used in by E2
 using UnitOfTime = std::chrono::nanoseconds;
 
-/// Strongly-typed unit of length used in this world. Defined here for future
-/// use; current aliases (Location, Distance) use plain float meters for simplicity.
-struct UnitOfLength {
-    int64_t _value = 0;
-
-    constexpr UnitOfLength() = default;
-
-    explicit constexpr UnitOfLength(int64_t value): _value(value) {}
-
-    constexpr UnitOfLength operator+(UnitOfLength o) const { return UnitOfLength(_value + o._value); }
-    constexpr UnitOfLength operator-(UnitOfLength o) const { return UnitOfLength(_value - o._value); }
-    constexpr UnitOfLength operator*(UnitOfLength o) const { return UnitOfLength(_value * o._value); }
-    constexpr UnitOfLength operator/(UnitOfLength o) const { return UnitOfLength(_value / o._value); }
-};
-
-/// Spacial location defined by 3D coordinate within the world.
-using WorldPosition = Vector3<UnitOfLength>;
-
-// Spacial orientation defined by a rotation primitive.
-using Rotation = glm::quat;
-
-// struct Transformation {
-//     glm::vec3 translation;
-//     glm::quat rotation;
-//     glm::vec3 scaling;
-// };
-
 typedef int64_t UniqueIdentifier;
 
 struct Universe;
